+++ davinci/branches/64bit/ff_ascii.c
@@ -49,10 +49,10 @@ ff_ascii(vfuncptr func, Var *arg)
   int     *idata = NULL;
   float   *fdata = NULL;
   double  *ddata = NULL;
-  int      count=0;
+  size_t   count=0;
   char    *delim = " \t";
   int      i,j,k;
-  int      dsize;
+  size_t   dsize;
   int      x=0;
   int      y=0;
   int      z=0;
